# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Go-Snip.",
        "title": "Go-Snip API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Nana Kwasi Asante",
            "url": "https://www.nkasante.com",
            "email": "asantekwasi101@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1",
    "paths": {
...
        "/snippets": {
            "get": {
                "description": "Get all snippets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snippets"
                ],
                "summary": "Get Snippets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.APISuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    }
                }
            }
        },
        "/snippets/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a Snippet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snippets"
                ],
                "summary": "Create a Snippet",
                "parameters": [
                    {
                        "description": "snippet",
                        "name": "Snippet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.NewSnippetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.APISuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    }
                }
            }
        },
        "/snippets/{id}": {
            "get": {
                "description": "Get a snippet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snippets"
                ],
                "summary": "Get Snippet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.APISuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Snippet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snippets"
                ],
                "summary": "Update Snippet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "snippet",
                        "name": "Snippet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.NewSnippetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.APISuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete snippet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Snippets"
                ],
                "summary": "Delete snippet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Snippet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.APISuccessMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.APIErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.APIErrorMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "types.APISuccessMessage": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "string"
                }
            }
        },
...
        "types.NewSnippetRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}